+++ src/app/api/books/route.ts
@@ -1,0 +1,8 @@
+import {BOOKS_DATA} from "@/utils/db.util";
+
+export async function GET(req: Request, res: Response) {
+    const body = JSON.stringify({results: [...BOOKS_DATA]});
+    return new Response(body, {
+        status: 200
+    })
+}