+++ src/app/layout.tsx
@@ -13,7 +13,7 @@ export const metadata = {
 export default function RootLayout({children}: { children: ReactNode }) {
     return (
         <html lang="en">
-        <body className={`${inter.className}`}>
+        <body className={`${inter.className} min-h-screen flex flex-col dark:bg-black`}>
         <Header/>
         <Content>{children}</Content>
         <DarkMode/>