+++ src/components/index.tsx
@@ -5,3 +5,7 @@ export {default as DarkMode} from "./layout/DarkMode";
 
 export {default as Container} from "./common/Container";
 export {default as ActiveLink} from "./common/ActiveLink";
+
+export {default as Carousel} from "./pages/Carousel";
+export {default as BookImage} from "./pages/BookImage";
+export {default as Book} from "./pages/Book";