+++ src/components/layout/Footer.tsx
@@ -1,9 +1,12 @@
 import React from 'react';
+import {Container} from "@/components";
 
 const Footer = () => {
     return (
-        <footer>
-            Footer
+        <footer className="mt-auto my-2">
+            <Container className="text-center">
+                <span className="text-red-600">❤</span>
+            </Container>
         </footer>
     );
 };