+++ src/services/books.service.ts
@@ -1,0 +1,5 @@
+import API from "@/config/API";
+
+export const getBooksService = () => {
+        return API.get<IResponse<ReadonlyArray<IBook>>>("http://localhost:3000/api/books");
+}