+++ src/types/books.d.ts
@@ -1,0 +1,12 @@
+interface IBook {
+    id: number | string;
+    name: string;
+    author: string;
+    summary: string;
+    description: string;
+    imageUrl: string;
+    stock: number;
+    rate: number;
+    price: number;
+    categories: string[];
+}