+++ src/types/common.d.ts
@@ -1,0 +1,3 @@
+interface IResponse<T> {
+    results: T;
+}